bits of changes
