+++ src/parse/parse.py
@@ -1,0 +1,22 @@
+import yaml
+import torch.distributions as D
+import torch
+
+from typing import Any, Callable, Optional
+
+from train import ContextTrainer, TrainerSteps
+from models import MODELS
+from function_classes import FUNCTION_CLASSES
+from core import ContextModel, FunctionClass
+from utils import curried_throw
+
+
+######################################
+###     Trainer Initialization     ###
+######################################
+
+
+######################################
+###  END: Trainer Initialization   ###
+######################################
+
